% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_display_plot.R
\name{mod_display_plot_server}
\alias{mod_display_plot_server}
\title{Plot variables Server module}
\usage{
mod_display_plot_server(id, var_inputs)
}
\arguments{
\item{id}{module id}

\item{var_inputs}{inputs from mod_var_input}
}
\value{
shiny server module
}
\description{
Plot variables Server module
}
