% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_display_plot.R
\name{mod_display_plot_ui}
\alias{mod_display_plot_ui}
\title{Plot variables UI module}
\usage{
mod_display_plot_ui(id)
}
\arguments{
\item{id}{module id}
}
\value{
shiny UI module
}
\description{
Plot variables UI module
}
