% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_var_input.R
\name{mod_var_input_server}
\alias{mod_var_input_server}
\title{Import variables Server module}
\usage{
mod_var_input_server(id)
}
\arguments{
\item{id}{module id}
}
\value{
shiny server module
}
\description{
Import variables Server module
}
